refactor: 불필요 코드 삭제
- contact-form.tsx 파일 삭제
- features.tsx: 사용하지 않는 의존성 삭제
- hero.tsx: 사용하지 않는 부분 삭제
