+++ client/app/recruit/page.tsx
@@ -5,11 +5,11 @@ import { Button } from "@/components/ui/button"
 export default function RecruitMenu() {
   return (
     <main className="container flex flex-col py-8">
-      <HeadingText subtext="워크넷 제공 구인 공고를 검색하실 수 있습니다">
+      <HeadingText subtext="사람인 제공 채용 정보를 검색하실 수 있습니다">
         채용정보
       </HeadingText>
       <section className="mt-6">
-        <div className="flex flex-row space-x-4 align-middle text-lg">
+        <div className="flex flex-row space-x-4 text-lg">
           <label htmlFor="title" className="pt-1 font-bold">
             제목
           </label>
