+++ client/components/pages/features.tsx
@@ -1,13 +1,12 @@
-import Image from "next/image"
 import HeadingText from "@/components/heading-text"
-import { features } from "@/config/contents"
+import { features } from "@/config/landing-content"
 import { Icons } from "@/components/icons"
 
 export default function Features() {
   return (
     <section className="container space-y-8 py-12 lg:py-20" id="features">
       {features.header || features.subheader ? (
-        <HeadingText subtext={features.subheader} className="text-center">
+        <HeadingText subtext={features.subheader}>
           {features.header}
         </HeadingText>
       ) : null}
